jump to __verbose_terminate_handler

+++ src/Middle/MyTasks.cpp
@@ -30,7 +30,7 @@ void prvAdis16488Task(void *pvParameters){
 		unsigned char tmp = 0;
 		unsigned char tmp2= 0;
 		SPI1Class::GetInstance()->ReadWrite(&tmp,&tmp2,1);
-		printf("%d\n\r",TIM_GetCounter(TIM1));
+		printf("%d\n\r",(int)TIM_GetCounter(TIM1));
 
 		vTaskDelay(1);
 	}
